+++ packages/vanilla/index.html
@@ -30,10 +30,10 @@
       <li><a href="src/basics/grid/tests/tests-grid.html">grid</a></li>
       <li><a href="src/basics/icon/tests/tests-icon.html">icon</a></li>
       <li><a href="src/components/icon-button/tests/tests-icon-button.html">icon button</a></li>
-      <li><a href="src/components/linear-progress-indicator/tests/test-linear-progress-indicator.html">linear progress indicator</a></li>
       <li><a href="src/components/modal/tests/tests-modal.html">modal</a></li>
       <li><a href="src/basics/form-elements/password-field/tests/tests-password-field.html">password field</a></li>
-      <li><a href="src/components/radial-progress-indicator/tests/test-radial-progress-indicator.html">radial progress indicator</a></li>
+      <li><a href="src/components/progress-bar/tests/test-progress-bar.html">progress bar</a></li>
+      <li><a href="src/components/progress-ring/tests/tests-progress-ring.html">progress ring</a></li>
       <li><a href="src/basics/form-elements/radio-button/tests/tests-radio-button.html">radio button</a></li>
       <li><a href="src/basics/form-elements/range/tests/tests-range.html">range</a></li>
       <li><a href="src/basics/rich-text/tests/tests-rich-text.html">rich text</a></li>
