+++ packages/vanilla/src/components/progress-bar/progress-bar.html
@@ -1,6 +1,6 @@
-<div class="hig__linear-progress-indicator">
-  <div class="hig__linear-progress-indicator__bar">
-    <div class="hig__linear-progress-indicator__fill"></div>
+<div class="hig__progress-bar">
+  <div class="hig__progress-bar__bar">
+    <div class="hig__progress-bar__fill"></div>
     <svg width="3px" height="4px" viewBox="0 0 3 4" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
       <polygon id="end-right" points="0 0 2.68 0 1 4 0 4" fill="#0ED3BE"></polygon>
     </svg>
