+++ packages/vanilla/src/components/progress-bar/tests/test-progress-bar.html
@@ -30,26 +30,26 @@
       });
       determinateHeader.mount(".header-determinate");
 
-      var determinate0 = new Hig.LinearProgressIndicator();
+      var determinate0 = new Hig.ProgressBar();
       determinate0.mount('.determinate-0');
       determinate0.setPercentComplete(0);
 
-      var determinate33 = new Hig.LinearProgressIndicator();
+      var determinate33 = new Hig.ProgressBar();
       determinate33.mount('.determinate-33');
       determinate33.setPercentComplete(33);
 
-      var determinate66 = new Hig.LinearProgressIndicator();
+      var determinate66 = new Hig.ProgressBar();
       determinate66.mount('.determinate-66');
       determinate66.setPercentComplete(66);
 
-      var determinate100 = new Hig.LinearProgressIndicator();
+      var determinate100 = new Hig.ProgressBar();
       determinate100.mount('.determinate-100');
       determinate100.setPercentComplete(100);
 
 
 
 
-      var animated = new Hig.LinearProgressIndicator();
+      var animated = new Hig.ProgressBar();
       animated.mount('.animated');
       animated.setPercentComplete(0);
 
@@ -73,7 +73,7 @@
       });
       indeterminateHeader.mount(".header-indeterminate");
 
-      var indeterminate = new Hig.LinearProgressIndicator();
+      var indeterminate = new Hig.ProgressBar();
       indeterminate.mount('.indeterminate');
     </script>
   </body>
