+++ packages/vanilla/src/components/progress-ring/progress-ring.html
