+++ packages/vanilla/src/components/progress-ring/tests/tests-progress-ring.html
@@ -30,26 +30,26 @@
       // });
       // determinateHeader.mount(".header-determinate");
 
-      // var determinate0 = new Hig.RadialProgressIndicator();
+      // var determinate0 = new Hig.ProgressRing();
       // determinate0.mount('.determinate-0');
       // determinate0.setPercentComplete(0);
 
-      // var determinate33 = new Hig.RadialProgressIndicator();
+      // var determinate33 = new Hig.ProgressRing();
       // determinate33.mount('.determinate-33');
       // determinate33.setPercentComplete(33);
 
-      // var determinate66 = new Hig.RadialProgressIndicator();
+      // var determinate66 = new Hig.ProgressRing();
       // determinate66.mount('.determinate-66');
       // determinate66.setPercentComplete(66);
 
-      // var determinate100 = new Hig.RadialProgressIndicator();
+      // var determinate100 = new Hig.ProgressRing();
       // determinate100.mount('.determinate-100');
       // determinate100.setPercentComplete(100);
 
 
 
 
-      // var animated = new Hig.RadialProgressIndicator();
+      // var animated = new Hig.ProgressRing();
       // animated.mount('.animated');
       // animated.setPercentComplete(0);
 
@@ -73,7 +73,7 @@
       });
       indeterminateHeader.mount(".header-indeterminate");
 
-      var indeterminate = new Hig.RadialProgressIndicator();
+      var indeterminate = new Hig.ProgressRing();
       indeterminate.mount('.indeterminate');
     </script>
   </body>
